[4. Median of Two Sorted Arrays][Accepted]committed by Anthony

+++ 4-Median-of-Two-Sorted-Arrays/solution.java
@@ -84,11 +84,11 @@ public class Solution{
         if (n % 2 == 1) {
             return findKthElement(nums1, 0, nums2, 0, n / 2 + 1);
         } else {
-            return (findKthElement(nums1, 0, nums2, 0, n / 2) + findKthElement(nums1, 0, nums2, 0, n / 2 + 1)) / 2.0;
+            return (findKthElement(nums1, 0, nums2, 0, n / 2) + findKthElement(nums1, 0, nums2, 0, n / 2 + 1)) / 2;
         }
     }
     
-    public int findKthElement(int[] nums1, int a, int[] nums2, int b, int k) {
+    public double findKthElement(int[] nums1, int a, int[] nums2, int b, int k) {
         if (a >= nums1.length) {
             return nums2[b + k - 1];
         }
